requirements.txt
baidu-aip==2.2.18.0
certifi==2021.10.8
charset-normalizer==2.0.9
Django==2.2
django-password-reset==2.0
idna==3.3
importlib-metadata==4.8.2
Markdown==3.3.6
Pillow==8.4.0
Pygments==2.10.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.7
wincertstore==0.2
zipp==3.6.0
